Extracted event data extractor to allow different serialization types for EventHub
