+++ src/main/java/com/microsoft/azure/eventhubs/kafka/connect/sink/EventHubSinkConfig.java
@@ -20,13 +20,26 @@ public class EventHubSinkConfig extends AbstractConfig {
      */
     public static final String CLIENTS_PER_TASK = "eventhub.clients.per.task";
 
+    /**
+     * EventHub serialization type for structured events with schema. By default all structured
+     * events are converted to JSON.
+     */
+    public static final String SERIALIZATION_TYPE = "eventhub.serialization";
+
+    /**
+     * Send all structured events to EventHub as JSON
+     */
+    public static final String SERIALIZATION_JSON = "json";
+
     private static final short defaultClientsPerTask = 1;
 
     public static final ConfigDef CONFIG_DEF = new ConfigDef()
             .define(CONNECTION_STRING, Type.STRING, Importance.HIGH,
                     "EventHub Connection String")
             .define(CLIENTS_PER_TASK, Type.SHORT, defaultClientsPerTask, Importance.HIGH,
-                    "Number of Event Hub clients to use per task");
+                    "Number of Event Hub clients to use per task")
+            .define(SERIALIZATION_TYPE, Type.STRING, SERIALIZATION_JSON, Importance.LOW,
+                    "Method of serialization of structured events with schema to EventHub");
 
     public EventHubSinkConfig(Map<String, String> configValues) {
         super(CONFIG_DEF, configValues);
